KB-229, KB230 Refactored CitationDocument out of curation task
Also, have begun intercepting PDF dissemination to use this.
Not working properly yet.
