+++ dspace-xmlui/dspace-xmlui-api/src/main/java/org/dspace/app/xmlui/cocoon/BitstreamReader.java
@@ -33,6 +33,7 @@ import org.dspace.core.ConfigurationManager;
 import org.dspace.core.Constants;
 import org.dspace.core.Context;
 import org.dspace.core.LogManager;
+import org.dspace.disseminate.CitationDocument;
 import org.dspace.handle.HandleManager;
 import org.dspace.usage.UsageEvent;
 import org.dspace.utils.DSpace;
@@ -40,9 +41,7 @@ import org.xml.sax.SAXException;
 
 import javax.mail.internet.MimeUtility;
 import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.UnsupportedEncodingException;
+import java.io.*;
 import java.net.URLEncoder;
 import java.sql.SQLException;
 import java.util.Map;
@@ -301,9 +300,64 @@ public class BitstreamReader extends AbstractReader implements Recyclable
             }
                 
             // Success, bitstream found and the user has access to read it.
-            // Store these for later retreval:
-            this.bitstreamInputStream = bitstream.retrieve();
-            this.bitstreamSize = bitstream.getSize();
+            // Store these for later retrieval:
+
+            //Due to the OSU Knowledge Bank policy of intercepting views to the original bitstream to instead show a
+            // citation altered version of the object, we need to check if this resource falls under the
+            // "show watermarked alternative" umbrella. At which time we will not return the "bitstream", but will
+            // instead on-the-fly generate the citation rendition.
+
+            // What will trigger a redirect/intercept?
+            // 1) Intercepting Enabled
+            // 2) This User is not an admin
+            // 3) This object is citation-able
+            boolean isCitationEnabled = true;
+            boolean isUserAdmin = AuthorizeManager.isAdmin(context);
+
+            CitationDocument citationDocument = new CitationDocument();
+
+            if (isCitationEnabled && !isUserAdmin && citationDocument.canGenerateCitationVersion(bitstream)) {
+                // on-the-fly citation generator
+                log.info(item.getHandle() + " - " + bitstream.getName() + " is citable.");
+                
+                File citedDocument = null;
+                FileInputStream fileInputStream = null;
+                
+                try {
+                    //Create the cited document
+                    citedDocument = citationDocument.makeCitedDocument(bitstream);
+                    if(citedDocument == null) {
+                        log.error("CitedDocument was null");
+                    } else {
+                        log.info("CitedDocument was ok," + citedDocument.getAbsolutePath());
+                    }
+                    
+                    
+                    fileInputStream = new FileInputStream(citedDocument);
+                    if(fileInputStream == null) {
+                        log.error("Error opening fileInputStream: ");
+                    } else {
+                        log.info("fileInputStream was ok");
+                    }
+                    
+                    this.bitstreamInputStream = fileInputStream;
+                    this.bitstreamSize = citedDocument.length();
+                    log.info("Got to here");
+                    
+                } catch (Exception e) {
+                    log.error("Caught an error with intercepting the citation document:" + e.getMessage());
+                }
+                
+                
+            } else {
+                log.info("Citation not gonna work for this bitstream. " + bitstream.getID());
+
+                this.bitstreamInputStream = bitstream.retrieve();
+                this.bitstreamSize = bitstream.getSize();
+            }
+
+            log.info("Citation: This Far.");
+
             this.bitstreamMimeType = bitstream.getFormat().getMIMEType();
             this.bitstreamName = bitstream.getName();
             if (context.getCurrentUser() == null)
